requirements.txt
aiohttp>=3.8.0
mcp-core>=0.1.0
duckduckgo-search>=3.9.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
pydantic>=2.0.0
ratelimit>=2.2.1
cachetools>=5.3.0
loguru>=0.7.0